requirements.txt
cffi==1.1.2
cryptography==1.0
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.7.9
django-admin-sortable==1.8.4
django-autoslug==1.7.2
django-classy-tags==0.6.2
django-cms==3.1.2
django-cms-named-menus==0.1.10
django-easy-select2==1.3
django-filer==0.9.12
django-mptt==0.7.4
django-polymorphic==0.7.1
django-reversion==1.8.5
django-sekizai==0.8.2
Django-Select2==4.3.1
django-toolbelt==0.0.1
django-treebeard==3.0
djangocms-admin-style==0.2.7
djangocms-column==1.5
djangocms-file==0.1
djangocms-flash==0.2.0
djangocms-googlemap==0.3
djangocms-inherit==0.1
djangocms-installer==0.7.9
djangocms-link==1.6.2
djangocms-picture==0.1
djangocms-style==1.5
djangocms-teaser==0.1
djangocms-text-ckeditor==2.6.0
djangocms-video==0.1
easy-thumbnails==2.2
enum34==1.0.4
gunicorn==19.3.0
html5lib==0.999
idna==2.0
ipaddress==1.0.14
jsonfield==1.0.3
MySQL-python==1.2.5
ndg-httpsclient==0.4.0
Pillow==2.9.0
psycopg2==2.6.1
pyasn1==0.1.8
pycparser==2.14
pyOpenSSL==0.15.1
pytz==2015.4
requests==2.7.0
six==1.9.0
static3==0.6.1
tzlocal==1.2
Unidecode==0.4.18
wheel==0.24.0
